moved or deleted unused blocks and components
