+++ client/src/components/Header.jsx
@@ -34,7 +34,6 @@ const Header = () => {
           {/* Conditional rendering of Login/Logout */}
           {user ? (
             <div className="flex items-center lg:order-2">
-              {/* <Notificatio^nModal /> */}
               <button>
                 <Bell />
               </button>
@@ -53,8 +52,6 @@ const Header = () => {
             </div>
           ) : (
             <div className="flex justify-center  items-center lg:order-2">
-              {/* <div className="text-gray-800 dark:text-white hover:bg-gray-50 focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-4 lg:px-5 py-2 lg:py-2.5 mr-2 dark:hover:bg-gray-700 focus:outline-none dark:focus:ring-gray-800"> */}
-              {/* <NavLink to="/login">Login</NavLink> */}
               <div>
                 <Login />
               </div>
@@ -64,7 +61,6 @@ const Header = () => {
               </div>
 
               {/* hamburger menu */}
-
               <button
                 data-collapse-toggle="mobile-menu2"
                 type="button"
