+++ client/src/components/Home.jsx
@@ -2,11 +2,7 @@ import React from 'react';
 import Hero from './Hero';
 import SectionTeam from './blocks/SectionTeam';
 import SectionFeature from './blocks/SectionFeature';
-import SectionContent from './blocks/SectionContent';
 import SectionTestimonials from './blocks/SectionTestimonials';
-import CookieInfoBanner from '../components/blocks/CookieInfoBanner';
-import Comments from '../pages/Comments';
-import Layout from "../pages/Layout";
 
 const Home = () => {
   return (
@@ -15,8 +11,6 @@ const Home = () => {
       <SectionFeature />
       <SectionTestimonials />
       <SectionTeam />
-      {/* <SectionContent /> */}
-      {/* <CookieInfoBanner /> */}
     </>
   );
 };
