+++ client/src/components/Main.jsx
@@ -1,9 +1,8 @@
 // import Notfound from "../components/blocks/Notfound";
+import React from 'react';
 import Comments from '../pages/Comments';
 import { AnimatePresence, motion } from 'framer-motion';
 import AnimatedPage from './AnimatedPage';
-import ProjMap from './ProjMap';
-import React from 'react';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import NewOrder from '../pages/NewOrder';
 import OrderDetails from '../pages/OrderDetails';
@@ -13,11 +12,9 @@ import Register from './Register';
 import ProtectedRoute from './ProtectedRoute';
 import Home from './Home';
 import Dashboard from '../pages/Dashboard';
-import Payment from '../pages/Payment';
-import Checkout from '../pages/Checkout';
+// import Payment from '../pages/Payment';
+// import Checkout from '../pages/Checkout';
 import Contact from '../pages/Contact';
-import Confirmation from '../components/Confirmation';
-import Layout from '../pages/Layout';
 
 const Main = () => {
   const location = useLocation();
@@ -43,11 +40,10 @@ const Main = () => {
               <Route path="/orders/:id/update" element={<UpdateOrder />} />
 
               {/* Payment */}
-              <Route path="/checkout" element={<Checkout />} />
-              <Route path="/payment" element={<Payment />} />
+              {/* <Route path="/checkout" element={<Checkout />} /> */}
+              {/* <Route path="/payment" element={<Payment />} /> */}
               <Route path="/comments" element={<Comments />} />
 
-              <Route path="/confirmation" element={<Confirmation />} />
             </Route>
 
             <Route path="/login" element={<LoginForm />} />
