+++ client/src/components/blocks/not used/CookieInfoBanner.jsx
