+++ client/src/components/blocks/not used/SectionContent.jsx
